feat: Remove struct definition of Vendor records from create_vendor and establish in a separate file
